SweetAlert

+++ src/app/components/user-card/user-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { User } from 'src/app/interfaces/user.interface';
 import { UsersService } from 'src/app/services/users.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-card',
@@ -19,18 +20,29 @@ export class UserCardComponent implements OnInit {
 
   deleteUser(pId : number | undefined){
     
-    if(pId !== undefined){
+    let result = confirm("Deseas borrar el usuario " + this.myUser.first_name + "?");
+    if(result){
+      if(pId !== undefined){
       
-      this.usersService.delete(pId)
-        .then( response => {
-          console.log(response)
-          if(response !== null){
-            alert('Usuario borrado correctamente');
-          }
-        })
+        this.usersService.delete(pId)
+          .then( response => {
+            console.log(response)
+            if(response.id){
+              Swal.fire(
+                'Usuario Eliminado',
+                'Has eliminado el usuario de forma correcta',
+                'success'
+              )
+            }else{
+              alert(response.console.error);
+            }
+          })
+  
+          .catch( err => console.log(err))
+        
+      }
 
-        .catch( err => console.log(err))
-      
     }
+    
   }
 }
